# Add option to preserve multiple underscores between words