# Description

This PR adds a configuration option to all `.*Namer` structs that implement this behavior in the spirit of helping other projects that depend on allowing multiple underscores when building Prometheus-compliant names.

This option is highly discouraged because it goes against the specification, so I'm not adding extra documentation.

## Warning!

During implementation, I noticed that LabelNamer was already preserving multiple underscores even without this new option. Therefore, this change can be considered a breaking change for LabelNamer 😬 

## Benchmark results

### BenchmarkBuild (MetricNamer)
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/otlptranslator
cpu: Apple M2 Pro
                                                                                     │ bench-BenchmarkBuild-main.txt │ bench-BenchmarkBuild-preserveunderscores.txt │
                                                                                     │            sec/op             │         sec/op          vs base              │
Build/withSuffixes=true/utf8Allowed=true/Basic_metric_with_no_special_characters-2                       49.29n ± 7%             47.68n ±  4%       ~ (p=0.132 n=6)
Build/withSuffixes=true/utf8Allowed=true/Counter_metric-2                                                72.61n ± 1%             74.97n ±  6%  +3.25% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=true/Gauge_ratio_metric-2                                            92.11n ± 1%             95.14n ±  1%  +3.29% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=true/Metric_with_per-unit_suffix_notation-2                          131.8n ± 0%             135.5n ±  1%  +2.73% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=true/Metric_with_special_characters-2                                48.62n ± 0%             49.89n ±  1%  +2.60% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=true/Metric_starting_with_digit-2                                    46.38n ± 1%             47.66n ±  1%  +2.77% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=true/Metric_with_multiple_underscores-2                              49.54n ± 1%             50.19n ±  0%  +1.30% (p=0.004 n=6)
Build/withSuffixes=true/utf8Allowed=true/Metric_with_complex_unit-2                                      172.2n ± 2%             177.1n ±  2%  +2.82% (p=0.009 n=6)
Build/withSuffixes=true/utf8Allowed=false/Basic_metric_with_no_special_characters-2                      203.6n ± 2%             207.0n ±  3%       ~ (p=0.145 n=6)
Build/withSuffixes=true/utf8Allowed=false/Counter_metric-2                                               258.2n ± 1%             262.5n ±  1%  +1.67% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=false/Gauge_ratio_metric-2                                           244.0n ± 0%             247.9n ±  1%  +1.58% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=false/Metric_with_per-unit_suffix_notation-2                         351.2n ± 1%             357.6n ±  1%  +1.81% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=false/Metric_with_special_characters-2                               259.9n ± 1%             268.7n ±  4%  +3.39% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=false/Metric_starting_with_digit-2                                   172.8n ± 0%             178.7n ±  1%  +3.39% (p=0.002 n=6)
Build/withSuffixes=true/utf8Allowed=false/Metric_with_multiple_underscores-2                             285.8n ± 4%             290.9n ±  1%       ~ (p=0.180 n=6)
Build/withSuffixes=true/utf8Allowed=false/Metric_with_complex_unit-2                                     486.9n ± 1%             511.9n ±  3%  +5.15% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=true/Basic_metric_with_no_special_characters-2                      27.41n ± 0%             28.12n ±  1%  +2.61% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=true/Counter_metric-2                                               27.43n ± 0%             27.96n ±  0%  +1.97% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=true/Gauge_ratio_metric-2                                           27.45n ± 1%             28.05n ±  0%  +2.19% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=true/Metric_with_per-unit_suffix_notation-2                         26.95n ± 0%             27.30n ±  0%  +1.30% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=true/Metric_with_special_characters-2                               29.29n ± 6%             29.73n ±  1%       ~ (p=0.061 n=6)
Build/withSuffixes=false/utf8Allowed=true/Metric_starting_with_digit-2                                   27.32n ± 2%             27.30n ±  7%       ~ (p=0.937 n=6)
Build/withSuffixes=false/utf8Allowed=true/Metric_with_multiple_underscores-2                             30.57n ± 5%             30.60n ± 26%       ~ (p=0.699 n=6)
Build/withSuffixes=false/utf8Allowed=true/Metric_with_complex_unit-2                                     27.53n ± 1%             28.36n ±  3%  +3.01% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=false/Basic_metric_with_no_special_characters-2                     143.6n ± 0%             138.1n ±  0%  -3.80% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=false/Counter_metric-2                                              145.7n ± 0%             140.2n ±  1%  -3.74% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=false/Gauge_ratio_metric-2                                          139.8n ± 1%             135.6n ±  0%  -3.04% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=false/Metric_with_per-unit_suffix_notation-2                        129.4n ± 1%             124.7n ±  0%  -3.63% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=false/Metric_with_special_characters-2                              230.2n ± 0%             228.0n ±  0%  -0.98% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=false/Metric_starting_with_digit-2                                  131.8n ± 1%             127.9n ±  1%  -3.03% (p=0.002 n=6)
Build/withSuffixes=false/utf8Allowed=false/Metric_with_multiple_underscores-2                            206.4n ± 3%             201.7n ±  3%       ~ (p=0.089 n=6)
Build/withSuffixes=false/utf8Allowed=false/Metric_with_complex_unit-2                                    138.2n ± 1%             135.3n ±  5%  -2.10% (p=0.050 n=6)
geomean                                                                                                  96.19n                  96.94n        +0.78%

                                                                                     │ bench-BenchmarkBuild-main.txt │ bench-BenchmarkBuild-preserveunderscores.txt │
                                                                                     │             B/op              │         B/op          vs base                │
Build/withSuffixes=true/utf8Allowed=true/Basic_metric_with_no_special_characters-2                        48.00 ± 0%             48.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Counter_metric-2                                                 96.00 ± 0%             96.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Gauge_ratio_metric-2                                             112.0 ± 0%             112.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_per-unit_suffix_notation-2                           120.0 ± 0%             120.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_special_characters-2                                 64.00 ± 0%             64.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_starting_with_digit-2                                     40.00 ± 0%             40.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_multiple_underscores-2                               80.00 ± 0%             80.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_complex_unit-2                                       176.0 ± 0%             176.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Basic_metric_with_no_special_characters-2                       160.0 ± 0%             160.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Counter_metric-2                                                288.0 ± 0%             288.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Gauge_ratio_metric-2                                            208.0 ± 0%             208.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_per-unit_suffix_notation-2                          256.0 ± 0%             256.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_special_characters-2                                240.0 ± 0%             240.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_starting_with_digit-2                                    120.0 ± 0%             120.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_multiple_underscores-2                              240.0 ± 0%             240.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_complex_unit-2                                      368.0 ± 0%             368.0 ± 0%       ~ (p=1.000 n=6)
Build/withSuffixes=false/utf8Allowed=true/Basic_metric_with_no_special_characters-2                       32.00 ± 0%             32.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Counter_metric-2                                                32.00 ± 0%             32.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Gauge_ratio_metric-2                                            32.00 ± 0%             32.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_per-unit_suffix_notation-2                          24.00 ± 0%             24.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_special_characters-2                                48.00 ± 0%             48.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_starting_with_digit-2                                    24.00 ± 0%             24.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_multiple_underscores-2                              64.00 ± 0%             64.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_complex_unit-2                                      32.00 ± 0%             32.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Basic_metric_with_no_special_characters-2                      64.00 ± 0%             64.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Counter_metric-2                                               64.00 ± 0%             64.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Gauge_ratio_metric-2                                           64.00 ± 0%             64.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_per-unit_suffix_notation-2                         56.00 ± 0%             56.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_special_characters-2                               160.0 ± 0%             160.0 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_starting_with_digit-2                                   56.00 ± 0%             56.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_multiple_underscores-2                             96.00 ± 0%             96.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_complex_unit-2                                     64.00 ± 0%             64.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                                                   82.46                  82.46       +0.00%
¹ all samples are equal

                                                                                     │ bench-BenchmarkBuild-main.txt │ bench-BenchmarkBuild-preserveunderscores.txt │
                                                                                     │           allocs/op           │      allocs/op        vs base                │
Build/withSuffixes=true/utf8Allowed=true/Basic_metric_with_no_special_characters-2                        2.000 ± 0%             2.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Counter_metric-2                                                 3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Gauge_ratio_metric-2                                             3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_per-unit_suffix_notation-2                           4.000 ± 0%             4.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_special_characters-2                                 2.000 ± 0%             2.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_starting_with_digit-2                                     2.000 ± 0%             2.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_multiple_underscores-2                               2.000 ± 0%             2.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=true/Metric_with_complex_unit-2                                       5.000 ± 0%             5.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Basic_metric_with_no_special_characters-2                       6.000 ± 0%             6.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Counter_metric-2                                                8.000 ± 0%             8.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Gauge_ratio_metric-2                                            7.000 ± 0%             7.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_per-unit_suffix_notation-2                          10.00 ± 0%             10.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_special_characters-2                                6.000 ± 0%             6.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_starting_with_digit-2                                    6.000 ± 0%             6.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_multiple_underscores-2                              6.000 ± 0%             6.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=true/utf8Allowed=false/Metric_with_complex_unit-2                                      12.00 ± 0%             12.00 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Basic_metric_with_no_special_characters-2                       1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Counter_metric-2                                                1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Gauge_ratio_metric-2                                            1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_per-unit_suffix_notation-2                          1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_special_characters-2                                1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_starting_with_digit-2                                    1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_multiple_underscores-2                              1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=true/Metric_with_complex_unit-2                                      1.000 ± 0%             1.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Basic_metric_with_no_special_characters-2                      3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Counter_metric-2                                               3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Gauge_ratio_metric-2                                           3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_per-unit_suffix_notation-2                         3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_special_characters-2                               4.000 ± 0%             4.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_starting_with_digit-2                                   3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_multiple_underscores-2                             3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
Build/withSuffixes=false/utf8Allowed=false/Metric_with_complex_unit-2                                     3.000 ± 0%             3.000 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                                                                   2.806                  2.806       +0.00%
¹ all samples are equal

```

### BenchmarkNormalizeLabel (LabelNamer)

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/otlptranslator
cpu: Apple M2 Pro
                                                   │ bench-BenchmarkNormalizeLabel-main.txt │ bench-BenchmarkNormalizeLabel-preserveunderscore.txt │
                                                   │                 sec/op                 │        sec/op         vs base               │
NormalizeLabel/empty_label-2                                                    80.06n ± 2%            81.44n ± 7%        ~ (p=0.093 n=6)
NormalizeLabel/label_with_colons-2                                              67.48n ± 2%            68.86n ± 1%   +2.03% (p=0.022 n=6)
NormalizeLabel/label_with_capital_letters-2                                     97.19n ± 2%            98.22n ± 0%        ~ (p=0.093 n=6)
NormalizeLabel/label_with_special_characters-2                                  91.41n ± 1%            92.86n ± 2%   +1.59% (p=0.004 n=6)
NormalizeLabel/label_with_foreign_characters-2                                  120.7n ± 0%            117.2n ± 2%   -2.86% (p=0.002 n=6)
NormalizeLabel/label_with_dots-2                                                62.22n ± 1%            65.33n ± 4%   +5.00% (p=0.002 n=6)
NormalizeLabel/label_starting_with_digits-2                                     62.65n ± 0%            65.22n ± 3%   +4.11% (p=0.002 n=6)
NormalizeLabel/label_starting_with_underscores-2                                135.8n ± 0%            139.4n ± 1%   +2.69% (p=0.002 n=6)
NormalizeLabel/label_starting_with_2_underscores-2                            118.250n ± 0%            5.666n ± 0%  -95.21% (p=0.002 n=6)
geomean                                                                         89.34n                 64.81n       -27.45%

…