@ArthurSens 2025-08-09
The codebase has changed a lot since then, it seemed easier to start from zero then rebasing that PR 😅