@aknuds1 2025-08-12
> The code here was pushed to https://github.com/prometheus/otlptranslator/pull/20!

Thanks! Can this PR be closed, to avoid confusion?